Re-structured the code. Still yet to get to compile

+++ arduino_bot/motor_bot.h
@@ -23,7 +23,7 @@ class MotorBot {
       back_right_motor(BACK_RIGHT_MOTOR),
       back_left_motor(BACK_LEFT_MOTOR) {}
 
-   void set_speed(uint8_t speed)
+   void set_speed(int speed)
    {
       front_right_motor.setSpeed(speed);
       front_left_motor.setSpeed(speed);
@@ -31,18 +31,18 @@ class MotorBot {
       back_left_motor.setSpeed(speed);
    }
 
-   void turn_left(uint8_t speed)
+   void turn_left(int speed)
    {
       set_speed(speed);
     
    }
 
-   void turn_right(uint8_t speed)
+   void turn_right(int speed)
    {
       set_speed(speed);
    }
 
-   void move_forward(uint8_t speed)
+   void move_forward(int speed)
    {
       set_speed(speed);
       front_right_motor.run(FORWARD);
